64-bit type changes and print format specifier cleanup; added suppressReplyNotifications to Calendar API; renamed newSubscriptionVideosFeedLink for happier static analysis; fixed double allocation in single-entry feeds

+++ Source/BaseClasses/GDataObject.h
@@ -580,7 +580,7 @@ forCategoryWithScheme:(NSString *)scheme
                         URI:(NSString *)attributeURI;
 
 - (NSXMLNode *)addToElement:(NSXMLElement *)element
-  attributeValueWithInteger:(int)val
+  attributeValueWithInteger:(NSInteger)val
                    withName:(NSString *)name;
 
 // adding child elements
